Update Storybook to 5.3
