--- .storybook/addons.js
@@ -1,6 +1,0 @@
-// The order of these imports defines the order of the tabs in the addons panel
-import '@storybook/addon-viewport/register';
-import '@storybook/addon-actions/register';
-import '@storybook/addon-notes/register';
-import '@storybook/addon-a11y/register';
-import '@storybook/addon-contexts/register';
