+++ package.json
@@ -102,14 +102,14 @@
     "@shopify/sewing-kit": "^0.112.0",
     "@shopify/shrink-ray": "^1.0.0",
     "@shopify/splash": "^0.0.8",
-    "@storybook/addon-a11y": "^5.2.4",
-    "@storybook/addon-actions": "^5.2.4",
+    "@storybook/addon-a11y": "^5.3.2",
+    "@storybook/addon-actions": "^5.3.2",
     "@storybook/addon-console": "^1.2.1",
-    "@storybook/addon-contexts": "^5.2.4",
-    "@storybook/addon-notes": "^5.2.4",
-    "@storybook/addon-viewport": "^5.2.4",
-    "@storybook/react": "^5.2.4",
-    "@storybook/theming": "^5.2.4",
+    "@storybook/addon-contexts": "^5.3.2",
+    "@storybook/addon-notes": "^5.3.2",
+    "@storybook/addon-viewport": "^5.3.2",
+    "@storybook/react": "^5.3.2",
+    "@storybook/theming": "^5.3.2",
     "@types/enzyme": "^3.10.3",
     "@types/enzyme-adapter-react-16": "^1.0.5",
     "@types/lodash": "^4.14.138",
