+++ playground/stories.tsx
@@ -4,13 +4,10 @@ import {DetailsPage} from './DetailsPage';
 
 // eslint-disable-next-line import/no-default-export, import/no-anonymous-default-export
 export default {
-  title: 'Playground|Playground',
+  title: 'Playground/Playground',
   parameters: {
     chromatic: {disable: true},
   },
 };
-
-(KitchenSink as any).story = {name: 'Kitchen sink'};
-(DetailsPage as any).story = {name: 'Details page'};
 
 export {Playground, KitchenSink, DetailsPage};
